replaced pointers used as unique ids by unic() calls; better wav handling that removes incorrect extra headers; some work on tabulated trigonometry; BLOCK_MEMMNGR disabled by default
git-svn-id: file:///srv/svn/kernel1/trunk@87 1e37f2df-8004-0410-ba96-adcf012cb36e

+++ uvariablename.cc
@@ -317,7 +317,7 @@ UVariableName::buildFullname(UCommand *command, UConnection *connection, bool wi
           snprintf(indexstr,fullnameMaxSize-strlen(name),"__%d",(int)e1->val);      
         else         
           if (e1->dataType == DATA_STRING) 
-            snprintf(indexstr,fullnameMaxSize-strlen(name),"__%s",(int)e1->str->str());
+            snprintf(indexstr,fullnameMaxSize-strlen(name),"__%s",e1->str->str());
           else {
             
             delete e1;
